+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1181** : Remove dropwizard config yaml from the docker images.
+
 ## [v6.0-beta.49] - 2019-06-24
 
 * Updated to allow stacks to be built at this version.