+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#1181** : Remove dropwizard config yaml from the docker images.
 
+* Issue **#1182** : Fixed use of `in folder` for data retention and receipt policies.
+
+
 ## [v6.0-beta.49] - 2019-06-24
 
 * Updated to allow stacks to be built at this version.